+++ CHANGELOG.md
@@ -8,6 +8,8 @@ To publish a new release run `scripts/release` from the project directory.
 
 ## [Unreleased]
 
+## [0.21.0] – 2020-08-12
+
 ### Added
 - Add `--paginate` flag to paginate the output of mdcat (see [GH-152]);
   with this flag mdcat sends it output to `less -R` or any alternative pager set in `$MDCAT_PAGER` or `$PAGER`.
@@ -408,4 +410,5 @@ To publish a new release run `scripts/release` from the project directory.
 [0.18.4]: https://github.com/lunaryorn/mdcat/compare/mdcat-0.18.3...mdcat-0.18.4
 [0.19.0]: https://github.com/lunaryorn/mdcat/compare/mdcat-0.18.4...mdcat-0.19.0
 [0.20.0]: https://github.com/lunaryorn/mdcat/compare/mdcat-0.19.0...mdcat-0.20.0
-[Unreleased]: https://github.com/lunaryorn/mdcat/compare/mdcat-0.20.0...HEAD
+[0.21.0]: https://github.com/lunaryorn/mdcat/compare/mdcat-0.20.0...mdcat-0.21.0
+[Unreleased]: https://github.com/lunaryorn/mdcat/compare/mdcat-0.21.0...HEAD